# backend/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEERON API",
    "description": "Precision Aquaculture Intelligence Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tanks": {
      "get": {
        "tags": [
          "Tanks"
        ],
        "summary": "List Tanks",
        "description": "Retrieves a list of fish tanks with pagination, optionally filtered by zone or structural type.",
        "operationId": "list_tanks_api_v1_tanks_get",
        "parameters": [
          {
            "name": "zone_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zone Id"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_TankResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tanks/{tank_id}": {
      "get": {
        "tags": [
          "Tanks"
        ],
        "summary": "Get Tank Detail",
        "description": "Retrieves static details and parameters for a specific tank.",
        "operationId": "get_tank_detail_api_v1_tanks__tank_id__get",
        "parameters": [
          {
            "name": "tank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tank Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_TankDetailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tanks/{tank_id}/environment": {
      "get": {
        "tags": [
          "Tanks"
        ],
        "summary": "Get Tank Environment",
        "description": "Retrieves the latest environmental metrics snapshot for the tank.",
        "operationId": "get_tank_environment_api_v1_tanks__tank_id__environment_get",
        "parameters": [
          {
            "name": "tank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tank Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_Union_TankEnvironmentSnapshotResponse__NoneType__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tanks/{tank_id}/predictions": {
      "get": {
        "tags": [
          "Tanks"
        ],
        "summary": "Get Tank Predictions",
        "description": "Retrieves the complete set of AI forecasts (PSI, disease, mortality, harvest) for the tank.",
        "operationId": "get_tank_predictions_api_v1_tanks__tank_id__predictions_get",
        "parameters": [
          {
            "name": "tank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tank Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_TankPredictionsResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/sensors": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get Sensor Status",
        "description": "Retrieves status indicators, operational state, and calibration due dates for all farm sensors.",
        "operationId": "get_sensor_status_api_v1_settings_sensors_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_SensorStatusResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/thresholds": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get Threshold Configs",
        "description": "Retrieves the standard warning/critical metric thresholds configured for the farm.",
        "operationId": "get_threshold_configs_api_v1_settings_thresholds_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_ThresholdConfigResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update Threshold Configs",
        "description": "Updates warning/critical threshold parameters for a specific metric.",
        "operationId": "update_threshold_configs_api_v1_settings_thresholds_put",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThresholdUpdatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_ThresholdConfigResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/insights/tanks/{tank_id}": {
      "get": {
        "tags": [
          "Insights"
        ],
        "summary": "Get Tank Insights",
        "description": "Computes and returns a list of transient AI insights for a specific tank.",
        "operationId": "get_tank_insights_api_v1_insights_tanks__tank_id__get",
        "parameters": [
          {
            "name": "tank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tank Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_List_AiInsightResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseResponse_List_AiInsightResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/AiInsightResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[List[AiInsightResponse]]"
      },
...
      "BaseResponse_List_SensorStatusResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SensorStatusResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[List[SensorStatusResponse]]"
      },
      "BaseResponse_List_ThresholdConfigResponse__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ThresholdConfigResponse"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[List[ThresholdConfigResponse]]"
      },
...
      "BaseResponse_TankDetailResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/TankDetailResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[TankDetailResponse]"
      },
...
      "BaseResponse_TankPredictionsResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/TankPredictionsResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[TankPredictionsResponse]"
      },
...
      "BaseResponse_ThresholdConfigResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "$ref": "#/components/schemas/ThresholdConfigResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[ThresholdConfigResponse]"
      },
      "BaseResponse_Union_TankEnvironmentSnapshotResponse__NoneType__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TankEnvironmentSnapshotResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "BaseResponse[Union[TankEnvironmentSnapshotResponse, NoneType]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_TankResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/TankResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "title": "PaginatedResponse[TankResponse]"
      },
      "PaginationMeta": {
        "properties": {
          "currentPage": {
            "type": "integer",
            "title": "Currentpage"
          },
          "totalPages": {
            "type": "integer",
            "title": "Totalpages"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "totalCount": {
            "type": "integer",
            "title": "Totalcount"
          }
        },
        "type": "object",
        "required": [
          "currentPage",
          "totalPages",
          "limit",
          "totalCount"
        ],
        "title": "PaginationMeta",
        "description": "Metadata carrying pagination attributes."
      },
...
      "SensorStatusResponse": {
        "properties": {
          "sensor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Sensor Id"
          },
          "hardware_id": {
            "type": "string",
            "title": "Hardware Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "calibration_due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Due At"
          },
          "tank_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tank Id"
          }
        },
        "type": "object",
        "required": [
          "sensor_id",
          "hardware_id",
          "type",
          "status"
        ],
        "title": "SensorStatusResponse"
      },
...
      "TankDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "volumeM3": {
            "type": "number",
            "title": "Volumem3"
          },
          "maxBiomassKg": {
            "type": "number",
            "title": "Maxbiomasskg"
          },
          "species": {
            "type": "string",
            "title": "Species"
          },
          "digitalTwinConfig": {
            "additionalProperties": true,
            "type": "object",
            "title": "Digitaltwinconfig"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "volumeM3",
          "maxBiomassKg",
          "species",
          "digitalTwinConfig",
          "created_at"
        ],
        "title": "TankDetailResponse"
      },
      "TankEnvironmentSnapshotResponse": {
        "properties": {
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "ph": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ph"
          },
          "dissolved_oxygen": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dissolved Oxygen"
          },
          "salinity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salinity"
          },
          "ammonia": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ammonia"
          },
          "turbidity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turbidity"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          }
        },
        "type": "object",
        "title": "TankEnvironmentSnapshotResponse"
      },
...
      "TankPredictionsResponse": {
        "properties": {
          "psi": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Psi"
          },
          "disease": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disease"
          },
          "mortality": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mortality"
          },
          "harvest": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harvest"
          }
        },
        "type": "object",
        "title": "TankPredictionsResponse"
      },
      "TankResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "type": {
            "type": "string",
            "maxLength": 50,
            "title": "Type"
          },
          "volumeM3": {
            "type": "number",
            "title": "Volumem3"
          },
          "maxBiomassKg": {
            "type": "number",
            "title": "Maxbiomasskg"
          },
          "species": {
            "type": "string",
            "maxLength": 100,
            "title": "Species"
          },
          "currentBiomassKg": {
            "type": "number",
            "title": "Currentbiomasskg"
          },
          "healthStatus": {
            "type": "string",
            "title": "Healthstatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "volumeM3",
          "maxBiomassKg",
          "species",
          "currentBiomassKg",
          "healthStatus"
        ],
        "title": "TankResponse"
      },
...
      "ThresholdUpdatePayload": {
        "properties": {
          "farmId": {
            "type": "string",
            "format": "uuid",
            "title": "Farmid"
          },
          "metricName": {
            "type": "string",
            "title": "Metricname"
          },
          "warningMin": {
            "type": "number",
            "title": "Warningmin"
          },
          "warningMax": {
            "type": "number",
            "title": "Warningmax"
          },
          "criticalMin": {
            "type": "number",
            "title": "Criticalmin"
          },
          "criticalMax": {
            "type": "number",
            "title": "Criticalmax"
          }
        },
        "type": "object",
        "required": [
          "farmId",
          "metricName",
          "warningMin",
          "warningMax",
          "criticalMin",
          "criticalMax"
        ],
        "title": "ThresholdUpdatePayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}